requirements.txt
django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt=5.3.1
mysqlclient==2.2.4
boto3==1.35.39
python-jenkins==1.8.2
django-cors-headers==4.5.0
drf-yasg==1.21.7
sqlalchemy
django-simpleui

